+++ plugins/video/render3d/software/soft_polyrender.cpp
@@ -35,7 +35,7 @@ csSoftPolygonRenderer::csSoftPolygonRenderer (
   SCF_CONSTRUCT_IBASE(0);
   csRefTrackerAccess::AddAlias (CS_STATIC_CAST(iRenderBuffer*, this), this);
   csSoftPolygonRenderer::parent = parent;
-  renderBufferNum = ~0;
+  renderBufferNum = (uint)~0;
   polysNum = 0;
 }
